requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
anyio==4.7.0
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
discord.py==2.4.0
exceptiongroup==1.2.2
frozenlist==1.5.0
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.1.0
propcache==0.2.1
pycparser==2.22
PyNaCl==1.5.0
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
spotipy==2.24.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.3
youtube-dl==2021.12.17
youtube-search-python==1.6.6
yt-dlp==2024.12.6
